+++ Services/Identity/Identity.API/Common/Mapping/MappingProfile.cs
@@ -4,12 +4,12 @@ using Identity.API.Models;
 namespace Identity.API.Common.Mapping
 {
     /// <summary>
-    /// Define Automapper profile for Account.API entities.
+    /// Define Automapper profile for Identity.API entities.
     /// </summary>
     public class MappingProfile : AutoMapper.Profile
     {
         /// <summary>
-        /// Constructor of Automapper profile for Account.API.
+        /// Constructor of Automapper profile.
         /// </summary>
         public MappingProfile()
         {
