+++ Services/Identity/Identity.API/Infrastructure/IdentityContext.cs
@@ -42,11 +42,5 @@ namespace Identity.API.Infrastructure
         {
             return base.Update(entity);
         }
-
-        /// <inheritdoc/>
-        public override EntityEntry Remove(object entity)
-        {
-            return base.Remove(entity);
-        }
     }
 }
