+++ Services/Identity/Identity.API/Startup.cs
@@ -48,7 +48,7 @@ namespace Identity.API
                 app.UseDeveloperExceptionPage();
             }
 
-            app.UseHttpsRedirection();
+            //app.UseHttpsRedirection();
             app.UseRouting();
 
             app.UseAuthentication();
